Fix audio resampler sign conversion errors and GPU class conflicts
- Fix 4 sign conversion errors in audio_resampler.cpp (lines 64, 66, 68, 98)
  - Add static_cast<size_t> for int/size_t arithmetic in sample calculations
- Remove unused variable 'age' in gpu_memory_optimizer.hpp (line 70)
- Resolve GPUMemoryOptimizer class redefinition in gpu_system_validation.cpp
  - Use forward declaration instead of conflicting headers
  - Provide minimal stub implementation to avoid header conflicts

Addresses Windows and Linux pre-build compilation failures.

+++ src/decode/src/audio_resampler.cpp
@@ -61,11 +61,11 @@ std::optional<AudioFrame> AudioResampler::resample(const AudioFrame& input) {
     if(in_fmt == AV_SAMPLE_FMT_NONE || out_fmt == AV_SAMPLE_FMT_NONE) return std::nullopt;
 
     if(input.format == SampleFormat::FLTP) {
-        in_samples = static_cast<int>(input.data.size() / (sizeof(float) * in_channels));
+        in_samples = static_cast<int>(input.data.size() / (sizeof(float) * static_cast<size_t>(in_channels)));
     } else if(input.format == SampleFormat::S16) {
-        in_samples = static_cast<int>(input.data.size() / (2 * in_channels));
+        in_samples = static_cast<int>(input.data.size() / (2 * static_cast<size_t>(in_channels)));
     } else if(input.format == SampleFormat::FLT) {
-        in_samples = static_cast<int>(input.data.size() / (sizeof(float) * in_channels));
+        in_samples = static_cast<int>(input.data.size() / (sizeof(float) * static_cast<size_t>(in_channels)));
     } else {
         return std::nullopt;
     }
@@ -95,7 +95,7 @@ std::optional<AudioFrame> AudioResampler::resample(const AudioFrame& input) {
     out.channels = out_channels;
     out.format = params_.out_format;
     size_t sample_size = (out.format == SampleFormat::S16) ? 2 : sizeof(float);
-    size_t total_bytes = static_cast<size_t>(converted) * out.channels * sample_size;
+    size_t total_bytes = static_cast<size_t>(converted) * static_cast<size_t>(out.channels) * sample_size;
     out.data.resize(total_bytes);
     if(out.format == SampleFormat::FLTP) {
         float* dst = reinterpret_cast<float*>(out.data.data());
